Fix [#15] #13에서 변경된 사항 conflict 난 부분 해결 및 리뷰 사항 일부 수정 작업 진행
