+++ morib/src/main/java/org/morib/server/api/homeViewApi/service/fetch/timer/FetchUserTimerServiceImpl.java
@@ -2,7 +2,7 @@ package org.morib.server.api.homeViewApi.service.fetch.timer;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.task.infra.TaskGateway;
-import org.morib.server.domain.timer.application.TimeManager;
+import org.morib.server.domain.timer.application.TimerManager;
 import org.morib.server.domain.timer.infra.TimerGateway;
 import org.springframework.stereotype.Service;
 
@@ -11,12 +11,12 @@ import org.springframework.stereotype.Service;
 public class FetchUserTimerServiceImpl implements FetchUserTimerService{
     private final TaskGateway taskGateway;
     private final TimerGateway timerGateway;
-    private final TimeManager timeManager;
+    private final TimerManager timerManager;
 
     @Override
     public void execute() {
         taskGateway.fetchTasksInRange();
         timerGateway.fetchTimer();
-        timeManager.aggregate();
+        timerManager.aggregate();
     }
 }
