+++ morib/src/main/java/org/morib/server/api/timerView/controller/TimerViewController.java
@@ -4,8 +4,7 @@ import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.morib.server.api.timerView.dto.StopTimerRequestDto;
 import org.morib.server.api.timerView.dto.TodoCardResponseDto;
-import org.morib.server.api.timerView.service.fetch.todo.FetchTodoFacade;
-import org.morib.server.api.timerView.service.stop.timer.StopTimerFacade;
+import org.morib.server.api.timerView.facade.TimerViewFacade;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,13 +20,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class TimerViewController {
 
-    private final StopTimerFacade stopTimerFacade;
-    private final FetchTodoFacade fetchTodoFacade;
+    private final TimerViewFacade timerViewFacade;
 
     @PostMapping("/timer/stop/{taskId}")
     public ResponseEntity<String> stopTimerAndFetchAccumulatedTime( // @AuthenticationPrincipal Long userId,
          @PathVariable Long taskId, @RequestBody StopTimerRequestDto dto){
-        stopTimerFacade.afterStopTimer(taskId, dto);
+        timerViewFacade.stopAfterSumElapsedTime(taskId, dto);
         return ResponseEntity.status(200).body("요청이 성공했습니다!");
     }
 
@@ -37,11 +35,8 @@ public class TimerViewController {
                                                             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate targetDate) {
         if(!targetDate.equals(LocalDate.now()))
             throw new IllegalArgumentException("오늘 당일의 날짜만 확인할 수 있습니다!");
-        return ResponseEntity.ok(fetchTodoFacade.fetchTodoCard(targetDate));
+        return ResponseEntity.ok(timerViewFacade.fetchTodoCard(targetDate));
     }
-
-
-
 
 
 }
