+++ morib/src/main/java/org/morib/server/api/timerView/dto/TodoCardResponseDto.java
@@ -2,5 +2,5 @@ package org.morib.server.api.timerView.dto;
 
 import java.util.List;
 
-public record TodoCardResponseDto(Long totalTimeOfToday,List<TaskInTodoCardDto> task) {
+public record TodoCardResponseDto(int totalTimeOfToday,List<TaskInTodoCardDto> task) {
 }
