--- morib/src/main/java/org/morib/server/api/timerView/facade/FetchTodoFacade.java
@@ -1,8 +1,0 @@
-package org.morib.server.api.timerView.facade;
-
-import java.time.LocalDate;
-import org.morib.server.api.timerView.dto.TodoCardResponseDto;
-
-public interface FetchTodoFacade {
-    TodoCardResponseDto fetchTodoCard(LocalDate targetDate);
-}
