+++ morib/src/main/java/org/morib/server/domain/task/application/FetchTaskServiceImpl.java
@@ -1,14 +1,22 @@
 package org.morib.server.domain.task.application;
 
+import java.time.LocalDate;
+import java.util.LinkedHashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.morib.server.domain.category.infra.Category;
 import org.morib.server.domain.task.infra.Task;
+import org.morib.server.domain.task.infra.TaskRepository;
+import org.morib.server.domain.timer.infra.Timer;
+import org.morib.server.domain.todo.infra.Todo;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class FetchTaskServiceImpl implements FetchTaskService {
 
-    private final TaskManager taskManager;
+    private final TaskRepository taskRepository;
 
     @Override
     public void fetch() {
@@ -18,9 +26,20 @@ public class FetchTaskServiceImpl implements FetchTaskService {
 
     @Override
     public Task fetchById(Long taskId) {
+        return taskRepository.findById(taskId).orElseThrow(() -> new IllegalArgumentException("해당 task가 없습니다."));
+    }
 
+    @Override
+    public LinkedHashSet<Task> fetchByTodoAndSameTargetDate(Todo todo, LocalDate targetDate) {
+        Set<Task> tasks = todo.getTasks();
+        return tasks.stream().
+            filter(task -> isTimerInTaskPresentTargetDate(task, targetDate))
+            .collect(Collectors.toCollection(LinkedHashSet::new));
+    }
 
-        return null;
+    private boolean isTimerInTaskPresentTargetDate(Task task, LocalDate targetDate) {
+        Set<Timer> timers = task.getTimers();
+        return timers.stream().anyMatch(t -> t.getTargetDate().equals(targetDate));
     }
