+++ morib/src/main/java/org/morib/server/domain/timer/application/TimerManager.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class TimeManager {
+public class TimerManager {
 
     public void aggregate() {
         // 각 task의 timer들을 종합해 오늘 나의 작업시간 계산
